Initial changes for the tax rate calculator project
Contains the vertical slice for the marginal tax api requested by the assignment.

The TotalIncomeTaxController contains a POST method, CalculateIncomeTax, that takes in a user's annual income and tax year

It utilizes the TaxService class to calculate the total income tax.

In turn the TaxService class uses the MarginalTaxClient to handle calling the tax bracket server and dealing with the responses

The test project contains the unit tests for the project